added FC tests
