+++ src/integration/test_suites/smoke_test_fc_pk_volatile_lock/smoke_test_fc_pk_volatile_lock.yml
@@ -1,0 +1,3 @@
+---
+seed: 1
+testname: smoke_test_fc_pk_volatile_lock
