+++ src/integration/test_suites/smoke_test_fc_prov_lcc_tokens/smoke_test_fc_prov_lcc_tokens.yml
@@ -1,0 +1,3 @@
+---
+seed: 1
+testname: smoke_test_fc_prov_lcc_tokens
