+++ src/integration/test_suites/smoke_test_fc_random_item_prov/smoke_test_fc_random_item_prov.yml
@@ -1,0 +1,3 @@
+---
+seed: 1
+testname: smoke_test_fc_random_item_prov
