+++ src/integration/test_suites/smoke_test_fc_registers/smoke_test_fc_registers.yml
@@ -1,0 +1,3 @@
+---
+seed: 1
+testname: smoke_test_fc_registers
