--- src/integration/test_suites/smoke_test_fuse_prov_with_lcc/smoke_test_fuse_prov_with_lcc.yml
@@ -1,3 +1,0 @@
----
-seed: 1
-testname: smoke_test_fuse_prov_with_lcc
